Functions

Functions

List of implemented Functions

Language_Fundamentals :

Sub Package

1 	Entering_Commands 	clcM

2 	Matrices_and_Arrays 	zerosM, onesM, randM, eyeM, diagM, catM, horzcatM, vertcatM, repelemM(limited), repmatM, linspaceM, logspaceM, lengthM, sizeM, numelM, isscalarM, isvectorM, transposeM, squeezeM

MatLang.blkdiagM

MatLang.catM

MatLang.clcM

MatLang.eyeM

MatLang.horzcatM

MatLang.isscalarM

MatLang.isvectorM

MatLang.lengthM

MatLang.linspaceM

MatLang.logspaceM

MatLang.ndimsM

MatLang.numelM

MatLang.onesM

MatLang.randM

MatLang.repelemM

MatLang.repmatM

MatLang.sizeM

MatLang.squeezeM

MatLang.transposeM

MatLang.vertcatM

MatLang.zerosM

To see the full progress see this link: https://github.com/juliamatlab/MatLang/projects

The full list of the functions that will be implemented in the future can be found Here

Functions Explanation and Examples

Language Fundamentals

Entering_Commands

MatLang.clcM — Method.

clcM()

Clear Command Window for Juno or Atom IDE

source

Matrices and Arrays

MatLang.blkdiagM — Method.

blkdiagM(A1,A2,...)

Creates a square matrix with A1,A2,... on the diagonal and the rest of the elements being 0. Works for both square and non-square matrices.

blkdiagM(A1,A2,..., :obj)

Returns the object itself if you want to use BlockDiagonals methods. use collect(obj) to get the array.

Examples

1 = 3 * ones(2, 2);
A2 = 4 * ones(2, 2);
A3 = rand(3, 3);
mBlkdiag1 = blkdiagM(A1, A2, A3)

mBlkdiag2 = blkdiagM(ones(2, 2), 2 * ones(2, 2)) # [ones(2, 2) zeros(2, 2); zeros(2, 2)  2 * ones(2, 2)]

A1 = ones(2, 4);
A2 = 2 * ones(3, 2);
mBlkdiag3 = blkdiagM(A1, A2) # [ones(2,4) zeros(2,2); zeros(3,4) 2*ones(3,2)]

mBlkdiag1obj = blkdiagM(:obj, A1, A2, A3)

mBlkdiag2obj = blkdiagM(:obj, ones(2, 2), 2 * ones(2, 2)) # Block Diagonal object for [ones(2, 2) zeros(2, 2); zeros(2, 2)  2 * ones(2, 2)]

source

MatLang.catM — Method.

catM(dim, A1, A2, ...)

concatenates A1, A2, … , An along dimension dim.

Examples

mCat1 = catM(1, ones(3, 3), zeros(3, 3)) # [ones(3, 3); zeros(3, 3)]

mCat2 = catM(2, ones(3, 3), zeros(3, 3)) # [ones(3, 3) zeros(3, 3)]

mCat3 = catM(1, ones(3, 3), zeros(3, 3), 2 * ones(3, 3)) # [ones(3, 3); zeros(3, 3); 2 * ones(3, 3)]

mCat4 = catM(3, ones(2, 2, 2), zeros(2, 2, 2)) # [ones(2, 2, 2) zeros(2, 2, 2)]

source

MatLang.eyeM — Method.

eyeM(dim)         # square dim*dim matrix
eyeM(Type, dim)    # square dim*dim matrix
eyeM(dim, like = anArray) # to make an array with similar type of anArray
eyeM(dim1, dim2)   # giving size as a separate input numbers
eyeM(Type, dim1, dim2)   # giving size as a separate input numbers
eyeM(dim1, dim2, like = anArray) # to make an array with similar type of anArray
eyeM(sizeAsTuple) # giving size as a Tuple
eyeM(Type, sizeAsTuple) # giving size as a Tuple
eyeM(sizeAsTuple, like = anArray) # to make an array with similar type of anArray
eyeM(sizeAsArray) # non-efficient Matlab way
eyeM(Type, sizeAsArray) # non-efficient Matlab way
eyeM(sizeAsArray, like = anArray) # to make an array with similar type of anArray

Creates 2D Identity matrix (can be non-square matrix).

eyeM is 2-dimensional by the definition, so you don't need to pass :mat argument for getting a 2-dimensional matrix from eyeM(dim)

Examples

mEye0 = eyeM(2); # [1 0 0; 0 1 0]

mEye1 = eyeM(2, 3); # [1 0 0; 0 1 0]

mEye2 = eyeM(Int32, 2, 3); # [1 0 0; 0 1 0]

mEye3 = eyeM(2, 3, like = zerosM(Int8, 2, 2))

mEye4 = eyeM((2, 2))  # giving size as a Tuple. In Julia we use (2,2) instead of giving it as [2,2]

mEye5 = eyeM(Int32, (2, 2))  # giving size as a Tuple. In Julia we use (2,2) instead of giving it as [2,2]

mEye6 = eyeM([2, 2]) # non-efficient Matlab way

mEye7 = eyeM(Int32, [2, 2]) # non-efficient Matlab way

s1 = size(ones(2, 3)) # getting size from another matrix or calculation
mEye8 = eyeM(s1)  # giving size as a variable (Tuple).

source

MatLang.horzcatM — Function.

horzcatM(A1, A2, …)

Concatenate arrays horizontally

Examples

mHCat1 = horzcatM(ones(3, 3), zeros(3, 3)) # [ones(3, 3) zeros(3, 3)]

source

MatLang.isscalarM — Method.

isscalarM(x)
isscalarM(:mat, x)

Returns boolean true if x is scalar.

It uses Broadcast.DefaultArrayStyle{0}, which basically are numbers (0-dimensional) and 1 dimensional-1 element number arrays.

To get a MATLAB way result, pass :mat argument. Doing this:

For arrays (of any element type), it calculates number of elements (using numelM).

For single strings it calculates number of characters (using numelM).

For others, if it is among {Number,Char,Bool}, then it is considered scalar.

Examples

bIsscalar1 = isscalarM(1) # true

bIsscalar2 = isscalarM(5 * ones(1, 1, 1)) # false

# bIsscalar3 = isscalarM("H") # gives error
#
# bIsscalar4 = isscalarM("Hi") # gives error

bIsscalar5 = isscalarM(["Hi"]) # false

bIsscalar6 = isscalarM(["Hi" "Bye"]) # false

bIsscalar7 = isscalarM('H') # true # becareful that in Julia, chars are always singular.

bIsscalar8 = isscalarM(true) # true

# Matlab way:
bIsscalarMat1 = isscalarM(:mat, 1) # true

bIsscalarMat2 = isscalarM(:mat, [1]) # true

bIsscalarMat3 = isscalarM(:mat, 5 * ones(1, 1, 1)) # true

bIsscalarMat4 = isscalarM(:mat, "H") # true

bIsscalarMat5 = isscalarM(:mat, "Hi") # false

bIsscalarMat6 = isscalarM(:mat, ["Hi"]) # true

bIsscalarMat7 = isscalarM(:mat, ["Hi" "Bye"]) # false

bIsscalarMat8 = isscalarM(:mat, 'H') # true # becareful that in Julia, chars are always singular.

bIsscalarMat9 = isscalarM(:mat, true) # true

source

MatLang.isvectorM — Method.

isvectorM(x)

Returns boolean true if x is a vector.

It uses AbstractVector, which basically are 1 dimensional arrays.

To get a MATLAB way result, pass :mat argument. Doing this:

For arrays (of any element type), it considers 1-dimensional arrays and also 2-dimensional arrays that one of their dimensions are singletone.

For others, if it is among {Number, AbstractString, Char, Bool}, then it is considered vectir.

Examples

A1 = [1; 2; 3] # or [1, 2, 3]
bIsvector1 = isvectorM(A1) # true

A2 = [1 2 3]
bIsvector2 = isvectorM(A2) # false

bIsvector3 = isvectorM(ones(3, 1)) # false

bIsvector4 = isvectorM(ones(1, 3)) # false

bIsvector5 = isvectorM(ones(3)) # true

bIsvector6 = isvectorM("Hi") # false

# Matlab Way:
bIsvector7 = isvectorM(["Hi", "Bye"]) # true

bIsvectorMat1 = isvectorM(:mat, A1) # true

bIsvectorMat2 = isvectorM(:mat, A2) # true

bIsvectorMat3 = isvectorM(:mat, ones(3, 1)) # true

bIsvectorMat4 = isvectorM(:mat, ones(1, 3)) # true

bIsvectorMat5 = isvectorM(:mat, ones(3)) # true

bIsvectorMat6 = isvectorM(:mat, "Hi") # true

bIsvectorMat7 = isvectorM(:mat, ["Hi", "Bye"]) # true

source

MatLang.lengthM — Method.

lengthM(A)

Length of the largest array dimension of A.

Examples

A1 = [2 3 4 5 6 7;
      1 2 3 4 5 6]
nLength1 = lengthM(A1); # 6

A2 = ["Hi" 6;
      "Goodbye" 5;
      "Hello" 1]
nLength2 = lengthM(A2); # 3

source

MatLang.linspaceM — Method.

linspaceM(start, stop)
linspaceM(start, stop, step)
linspaceM(:arr, start, stop)
linspaceM(:arr, start, stop, step)

Generate linearly spaced range. You can also write this as start:stop or start:step:stop.

To get the full vector isntead of a range object pass :arr as the 1st argument.

Examples

mLinspace1 = linspaceM(1, 10) # 1:10

mLinspace2 = linspaceM(1, 10, 2) # 1:2:10 == 1:2:9

mLinspace3 = linspaceM(:arr, 1, 10) # collect(1:10)

mLinspace4 = linspaceM(:arr, 1, 10, 2) # collect(1:2:10)

source

MatLang.logspaceM — Method.

logspaceM(start, stop) # gives 50 points
logspaceM(start, stop, length)
logspaceM(start, stop, :equal) # equally spaced powers of 10

Generate logarithmically spaced vector (between 10^start and 10^stop). length is the number of the poinst (50 by defualt if not supplied). If you instead provide `:equal, it makes a range from equally spaced powers of 10

For logspaceM in contrast to linspaceM, the full vector is given by default.

Examples

mLogspace1 = logspaceM(1, 5) # 50 logarithmically spaced points between 10^1 and 10^5

mLogspace2 = logspaceM(1, 5, 10) # 10 logarithmically spaced points between 10^1 and 10^5

mLogspace3 = logspaceM(1, 5, :equal) # [10.0^1, 10.0^2, 10.0^3, 10.0^4, 10.0^5] == 10.0.^(1:5)

source

MatLang.ndimsM — Function.

ndimsM(A)

Return the number of dimensions of A.

In contrast to Matlab equivalent, this function considers trailing singleton dimensions,

Consider the points that are explained here when using this function: https://juliamatlab.github.io/MatLang/dev/juliavsmatlab/#Julia-Arrays:-1

Examples

A = rand(3,4,5);
nDim1 = ndims(A) # 3

source

MatLang.numelM — Function.

numelM(A)

Returns the number of elements of array A.

Examples

nNumel1 = numelM(1:5) # 5

nNumel2 = numelM([1, 2, 3, 4]) # 4

nNumel3 = numelM([1 2; 3 4]) # 4

source

MatLang.onesM — Method.

onesM(:mat, dim)         # square dim*dim matrix
onesM(:mat, Type,dim)    # square dim*dim matrix
onesM(:mat, dim, like=anArray) # to make an array with similar type of anArray
onesM(sizeAsArray) # non-efficient Matlab way
onesM(Type, sizeAsArray) # non-efficient Matlab way
onesM(sizeAsArray, like=anArray) # to make an array with similar type of anArray

returns an array filled with one values.

onesM(dim) returns 1-dimensional array. To get a square matrix like in Matlab, pass :mat as the 1st argument.

Examples

mOnes0 = onesM(:mat, 2) # same as ones(2,2)

mOnes1 = onesM(:mat, 2, like = zerosM(Int32, 2, 2)) # like method

mOnes2 = onesM(2) # same as ones(2)

mOnes3 = onesM(Int32, 2, 2) # same as ones(Int32,2,2)

mOnes4 = onesM((2, 2)) # = onesM(2,2) # giving size as Tuple

mOnes5 = onesM(Int32, (2, 2)) # giving size as Tuple and also the Int32 type

mOnes6 = onesM([2, 2]) # giving size as an Array, non-efficient Matlab way. Array should be Integer otherwise you will get errors.

mOnes7 = onesM(Int32, [2, 2]) # giving size as Array, non-efficient Matlab way

source

MatLang.randM — Method.

randM(:mat, dim)         # square dim*dim matrix
randM(:mat, Type, dim)    # square dim*dim matrix
randM(:mat, dim, like = anArray) # to make an array with similar type of anArray
randM(sizeAsArray) # non-efficient Matlab way
randM(Type, sizeAsArray) # non-efficient Matlab way
randM(sizeAsArray, like = anArray) # to make an array with similar type of anArray

Examples

mRand0 = randM(:mat, 2) # same as rand(2,2)

mRand1 = randM(:mat, 2, like = zerosM(Int32, 2, 2)) # like method

mRand2 = randM(2) # same as rand(2)

mRand3 = randM(Int32, 2, 2) # same as rand(Int32,2,2)

mRand4 = randM((2, 2)) # = onesM(2,2) # giving size as Tuple

mRand5 = randM(Int32, (2, 2)) # giving size as Tuple and also the Int32 type

mRand6 = randM([2, 2]) # giving size as an Array, non-efficient Matlab way. Array should be Integer otherwise you will get errors.

mRand7 = randM(Int32, [2, 2]) # giving size as Array, non-efficient Matlab way

source

MatLang.repelemM — Method.

repelemM(V, count)

Construct an array by repeating elements of array V by a given number of times specified by counts. if If count is a scalar, then each element of v is repeated "count" times

Examples

V1 = [1 2 3 4];
mRepelem1 = repelemM(V1, 3) # [1 1 1 2 2 2 3 3 3 4 4 4]

source

MatLang.repmatM — Method.

repmatM(A, n) # n*n copy
repmatM(A, s1, s2, ...)
repmatM(A, [s1, s2, ...])
repmatM(A, (s1, s2, ...))

Repeat copies of array A based on the give size

Examples

mRempat1 = repmatM(10, 3, 2) # [10 10; 10 10; 10 10]

V = [1; 2]
mRempat2 = repmatM(V, 3, 2)   # [[1; 2] [1; 2]; [1; 2] [1; 2]; [1; 2] [1; 2]]

mRempat3 = repmatM(V, [3, 2]) # [[1; 2] [1; 2]; [1; 2] [1; 2]; [1; 2] [1; 2]]

mRempat4 = repmatM(V, (3, 2)) # [[1; 2] [1; 2]; [1; 2] [1; 2]; [1; 2] [1; 2]]

mRempat5 = repmatM(V, 2) # [[1; 2] [1; 2]; [1; 2] [1; 2]]

mRempat6 = repmatM(V, 1, 3) # [[1; 2] [1; 2] [1; 2]]

source

MatLang.sizeM — Method.

sizeAsTuple = sizeM(A)
sizeDim = sizeM(dim)  # only returns the specified dim
sizeDimAsTuple = sizeM(A, dim1, dim2,...) # only returns the specified dims
sizeAsArray = sizeM(:arr, A)
sizeDimAsArray = sizeM(:arr, A, dim1, dim2,...) # only returns the specified dims
sz1, sz2, ... = size(A)

Returns the size of an array as a Tuple.

Pass :arr to get size as an Array (not a Tuple).

Consider the points that are explained here when using this function: https://juliamatlab.github.io/MatLang/dev/juliavsmatlab/#Julia-Arrays:-1

Examples

A1 = [2 3 4 5 6 7;
      1 2 3 4 5 6]
mSize1 = sizeM(:arr, A1); # [2; 6]
tSize1 = sizeM(A1); # (2, 6)

nA1Size2 = sizeM(A1, 2); # 6
sizeM(A1, 2) == 6 # true
sizeM(:arr, A1, 2) == 6 # false
sizeM(:arr, A1, 2) == [6] # true

nA1Size1, nA1Size2 = sizeM(A1); # 2 and 6

A2 = rand(3, 5, 4)
mSize2 = sizeM(:arr, A2, 2, 3); # [5; 4]
tSize2 = sizeM(A2, 2, 3); # (5, 4)

source

MatLang.squeezeM — Method.

squeezeM(A)

Drops all of the singleton dimensions of A (dimensions that are 1). If A contains only one element (i.e., all of its dimensions are singletons) then the output will be a zero-dimensional array.

If you know the dimension that you want to drop, use dropdims(A ; dims= dimensionsToRemove).

Only use this function if you don't know the dimensions that you want to remove, and you are sure that you are not removing important dimensions, and if you don't care about type stability.

Returns an array containing the same data as A but with no singleton dimensions; note that arr is NOT a copy of A, i.e., modifying the contents of arr will modify the contents of A. To get a copy use copy(arr).

Examples

A1 = ones(2, 1, 2); # 3 dimensional
mSqueeze1 = squeezeM(A1) # [1 1; 1 1]

A2 = zeros(1, 4, 1);
A2[:, 1:4, ] = [5; 3; 6; 0]
mSqueeze2 = squeezeM(A2) # When it gets one dimensional, it is vertical.

mSqueeze2 == [5; 3; 6; 0] # true
mSqueeze2 == [5 3 6 0] # false

source

MatLang.transposeM — Method.

transposeM(A)
transposeM(:arr, A)

Returns the transpose an array.

If :arr is supplied permutedims method is used which returns an array rather a transpose object.

Becareful that ' in Julia is used for adjoint and .' is not defined.

Examples

A1 = [2 3 4 5 6 7;
      1 2 3 4 5 6]
mTranspose1 = transposeM(A1)

mTranspose2 = transposeM(:arr, A1)

source

MatLang.vertcatM — Function.

vertcatM(A1, A2, …)

Concatenate arrays vertically

Examples

mVCat1 = vertcatM(ones(3, 3), zeros(3, 3)) # [ones(3, 3); zeros(3, 3)]

source

MatLang.zerosM — Method.

zerosM(:mat, dim)         # square dim*dim matrix
zerosM(:mat, Type,dim)    # square dim*dim matrix
zerosM(:mat, dim, like=anArray) # to make an array with similar type of anArray
zerosM(sizeAsArray) # non-efficient Matlab way
zerosM(Type, sizeAsArray) # non-efficient Matlab way
zerosM(sizeAsArray, like=anArray) # to make an array with similar type of anArray

returns an array filled with zero values.

zerosM(dim) returns 1-dimensional array. To get a square matrix like in Matlab, pass :mat as the 1st argument.

Examples

mZeros0 = zerosM(:mat, 2) # same as zeros(2,2)

mZeros1 = zerosM(:mat, 2, like = zerosM(Int32, 2, 2)) # like method

mZeros2 = zerosM(2) # same as zeros(2)

mZeros3 = zerosM(Int32, 2, 2) # same as zeros(Int32,2,2)

mZeros4 = zerosM((2, 2)) # = zerosM(2,2) # giving size as Tuple

mZeros5 = zerosM(Int32, (2, 2)) # giving size as Tuple and also the Int32 type

mZeros6 = zerosM([2, 2]) # giving size as an Array, non-efficient Matlab way. Array should be Integer otherwise you will get errors.

mZeros7 = zerosM(Int32, [2, 2]) # giving size as Array, non-efficient Matlab way

source