Functions

Functions

List of implemented Functions

Language_Fundamentals :

Sub Package

1 	Entering_Commands 	clcM

2 	Matrices_and_Arrays 	zerosM, onesM, randM, eyeM, diagM, catM, horzcatM, vertcatM, squeezeM

MatLang.blkdiagM

MatLang.catM

MatLang.clcM

MatLang.eyeM

MatLang.horzcatM

MatLang.onesM

MatLang.randM

MatLang.repelemM

MatLang.squeezeM

MatLang.vertcatM

MatLang.zerosM

To see the full progress see this link: https://github.com/juliamatlab/MatLang/projects

The full list of the functions that will be implemented in the future can be found Here

Functions Explanation and Examples

Language Fundamentals

Entering_Commands

MatLang.clcM — Method.

clcM()

Clear Command Window for Juno or Atom IDE

source

Matrices and Arrays

MatLang.blkdiagM — Method.

blkdiagM(A1,A2,...)

Creates a square matrix with A1,A2,... on the diagonal and the rest of the elements being 0. Works for both square and non-square matrices.

blkdiagM(A1,A2,..., :obj)

Returns the object itself if you want to use BlockDiagonals methods. use collect(obj) to get the array.

Examples

1 = 3 * ones(2, 2);
A2 = 4 * ones(2, 2);
A3 = rand(3, 3);
mBlkdiag1 = blkdiagM(A1, A2, A3)

mBlkdiag2 = blkdiagM(ones(2, 2), 2 * ones(2, 2)) # [ones(2, 2) zeros(2, 2); zeros(2, 2)  2 * ones(2, 2)]

A1 = ones(2, 4);
A2 = 2 * ones(3, 2);
mBlkdiag3 = blkdiagM(A1, A2) # [ones(2,4) zeros(2,2); zeros(3,4) 2*ones(3,2)]

mBlkdiag1obj = blkdiagM(:obj, A1, A2, A3)

mBlkdiag2obj = blkdiagM(:obj, ones(2, 2), 2 * ones(2, 2)) # Block Diagonal object for [ones(2, 2) zeros(2, 2); zeros(2, 2)  2 * ones(2, 2)]

source

MatLang.catM — Method.

catM(dim, A1, A2, ...)

concatenates A1, A2, … , An along dimension dim.

Examples

mCat1 = catM(1, ones(3, 3), zeros(3, 3)) # [ones(3, 3); zeros(3, 3)]

mCat2 = catM(2, ones(3, 3), zeros(3, 3)) # [ones(3, 3) zeros(3, 3)]

mCat3 = catM(1, ones(3, 3), zeros(3, 3), 2 * ones(3, 3)) # [ones(3, 3); zeros(3, 3); 2 * ones(3, 3)]

mCat4 = catM(3, ones(2, 2, 2), zeros(2, 2, 2)) # [ones(2, 2, 2) zeros(2, 2, 2)]

source

MatLang.eyeM — Method.

eyeM(dim)         # square dim*dim matrix
eyeM(Type, dim)    # square dim*dim matrix
eyeM(dim, like = anArray) # to make an array with similar type of anArray
eyeM(dim1, dim2)   # giving size as a separate input numbers
eyeM(Type, dim1, dim2)   # giving size as a separate input numbers
eyeM(dim1, dim2, like = anArray) # to make an array with similar type of anArray
eyeM(sizeAsTuple) # giving size as a Tuple
eyeM(Type, sizeAsTuple) # giving size as a Tuple
eyeM(sizeAsTuple, like = anArray) # to make an array with similar type of anArray
eyeM(sizeAsArray) # non-efficient Matlab way
eyeM(Type, sizeAsArray) # non-efficient Matlab way
eyeM(sizeAsArray, like = anArray) # to make an array with similar type of anArray

Creates 2D Identity matrix (can be non-square matrix).

Only 2 dimensions should be supplied othetwise you will receive an error.

Examples

mEye0=eyeM(2); # [1 0 0; 0 1 0]

mEye1=eyeM(2,3); # [1 0 0; 0 1 0]

mEye2=eyeM(Int32,2,3); # [1 0 0; 0 1 0]

mEye3=eyeM(2,3, like = zerosM(Int8,2,2) )

mEye4=eyeM((2,2))  # giving size as a Tuple. In Julia we use (2,2) instead of giving it as [2,2]

mEye5=eyeM(Int32,(2,2))  # giving size as a Tuple. In Julia we use (2,2) instead of giving it as [2,2]

mEye6=eyeM([2,2]) # non-efficient Matlab way

mEye7=eyeM(Int32,[2,2]) # non-efficient Matlab way

s1=size(ones(2,3)) # getting size from another matrix or calculation
mEye8=eyeM(s1)  # giving size as a variable (Tuple).

source

MatLang.horzcatM — Function.

horzcatM(A1, A2, …)

Concatenate arrays horizontally

Examples

mHCat1 = horzcatM(ones(3, 3), zeros(3, 3)) # [ones(3, 3) zeros(3, 3)]

source

MatLang.onesM — Method.

onesM(dim)         # square dim*dim matrix
onesM(Type, dim)    # square dim*dim matrix
onesM(dim, like = anArray) # to make an array with similar type of anArray
onesM(sizeAsArray) # non-efficient Matlab way
onesM(Type, sizeAsArray) # non-efficient Matlab way
onesM(sizeAsArray, like = anArray) # to make an array with similar type of anArray

Examples

mOnes0=onesM(2) # same as ones(2,2)

mOnes1 = onesM(2, like = zerosM(Int32,2,2) )

mOnes2=onesM(2,2) # same as ones(2,2)

mOnes3=onesM(Int32,2,2) # same as ones(Int32,2,2)

mOnes4=onesM((2,2)) # giving size as Tuple

mOnes5=onesM(Int32,(2,2)) # giving size as Tuple and also the Int32 type

mOnes6=onesM([2,2]) # giving size as an Array, non-efficient Matlab way. Array should be Integer otherwise you will get errors.

mOnes7=onesM(Int32,[2,2]) # giving size as Array, non-efficient Matlab way

source

MatLang.randM — Method.

randM(dim)         # square dim*dim matrix
randM(Type, dim)    # square dim*dim matrix
randM(dim, like = anArray) # to make an array with similar type of anArray
randM(sizeAsArray) # non-efficient Matlab way
randM(Type, sizeAsArray) # non-efficient Matlab way
randM(sizeAsArray, like = anArray) # to make an array with similar type of anArray

Examples

mRand0=randM(2) # same as rand(2,2)

mRand1 = randM(2, like = zerosM(Int32,2,2) )

mRand2=randM(2,2) # same as rand(2,2)

mRand3=randM(Int32,2,2) # same as rand(Int32,2,2)

mRand4=randM((2,2)) # giving size as Tuple

mRand5=randM(Int32,(2,2)) # giving size as Tuple and also the Int32 type

mRand6=randM([2,2]) # giving size as an Array, non-efficient Matlab way. Array should be Integer otherwise you will get errors.

mRand7=randM(Int32,[2,2]) # giving size as Array, non-efficient Matlab way

source

MatLang.repelemM — Method.

repelemM(V, count)

Construct an array by repeating elements of array V by a given number of times specified by counts. if If count is a scalar, then each element of v is repeated "count" times

Examples

V1 = [1 2 3 4];
mRepelem1 = repelemM(V1, 3) # [1 1 1 2 2 2 3 3 3 4 4 4]

source

MatLang.squeezeM — Method.

squeezeM(A)

Drops all of the singleton dimensions of A (dimensions that are 1). If A contains only one element (i.e., all of its dimensions are singletons) then the output will be a zero-dimensional array.

If you know the dimension that you want to drop, use dropdims(A ; dims= dimensionsToRemove).

Only use this function if you don't know the dimensions that you want to remove, and you are sure that you are not removing important dimensions, and if you don't care about type stability.

Returns an array containing the same data as A but with no singleton dimensions; note that arr is NOT a copy of A, i.e., modifying the contents of arr will modify the contents of A. To get a copy use copy(arr).

Examples

A1 = ones(2, 1, 2); # 3 dimensional
mSqueeze1 = squeezeM(A1) # [1 1; 1 1]

A2 = zeros(1, 4, 1);
A2[:, 1:4, ] = [5; 3; 6; 0]
mSqueeze2 = squeezeM(A2) # When it gets one dimensional, it is vertical.

mSqueeze2 == [5; 3; 6; 0] # true
mSqueeze2 == [5 3 6 0] # false

source

MatLang.vertcatM — Function.

vertcatM(A1, A2, …)

Concatenate arrays vertically

Examples

mVCat1 = vertcatM(ones(3, 3), zeros(3, 3)) # [ones(3, 3); zeros(3, 3)]

source

MatLang.zerosM — Method.

zerosM(dim)         # square dim*dim matrix
zerosM(Type,dim)    # square dim*dim matrix
zerosM(dim, like=anArray) # to make an array with similar type of anArray
zerosM(sizeAsArray) # non-efficient Matlab way
zerosM(Type, sizeAsArray) # non-efficient Matlab way
zerosM(sizeAsArray, like=anArray) # to make an array with similar type of anArray

Examples

mZeros0=zerosM(2) # same as zeros(2,2)

mZeros1 = zerosM(2, like = zerosM(Int32,2,2) ) # like method

mZeros2=zerosM(2,2) # same as zeros(2,2)

mZeros3=zerosM(Int32,2,2) # same as zeros(Int32,2,2)

mZeros4=zerosM((2,2)) # giving size as Tuple

mZeros5=zerosM(Int32,(2,2)) # giving size as Tuple and also the Int32 type

mZeros6=zerosM([2,2]) # giving size as an Array, non-efficient Matlab way. Array should be Integer otherwise you will get errors.

mZeros7=zerosM(Int32,[2,2]) # giving size as Array, non-efficient Matlab way

source